requirements.txt
Flask==3.1.3
openai==1.109.1
python-dotenv==1.2.2
